Status tracking in jobs.py!
This commit adds a basic status tracking system for any script-based
job submission. At job creation time an additional (empty) file named
the same as the job script script but with the extension
.status.[STATUS] instead of .sh will be created, with [STATUS] being
updated through various stages either by the job submission script or
the job script itself:

  - created: job script was created
  - submitted: script was submitted (currently all jobs will move from
    created to submitted near-instantly, even if the submission itself
    failed - will try to detect submission failure in the future)
  - running: set by the job script itself at the start
  - done: set by the job script if everything finished ok
  - exit.[EXIT_CODE]: set by the job script if any error occured - the
    exit code from the part that failed is added at the end

Note this status tracking system has to be enabled explicitly by adding
--tracking. Repeating the submission command will cause jobs.py to first
check if any .status file exists for each job, and if so skips the job
submission step. The user can currently remove any .status files by
hand, and repeat the submission command to effectively resubmit those
jobs. If '--tracking long' is specified a list of the scripts in each
status will also be printed at the end, otherwise only a short summary.

NB: it's possible a batch job could fail or be killed in some way that
leaves the status as 'running'. Will investigate if it's possible to
figure out if a job is actually running or not by interacting with the
batch system directly.